backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg
pillow
numpy
trimesh
torch
torchvision
onnxruntime
transformers
scipy
accelerate
einops
omegaconf
opencv-python
depth-anything-v2
